+++ printlinter/__init__.py
@@ -2,13 +2,20 @@
 """Print-Linter project's code."""
 
 # Local imports
-from .classes import IgnoreFile, IgnoreLine, IssueEnum, IssueInfo  # noqa: F401
+from .classes import (  # noqa: F401
+    IgnoredBlock,
+    IgnoreFile,
+    IgnoreLine,
+    IssueEnum,
+    IssueInfo,
+)
 from .config import DEFAULT_IGNORED_REP, Config  # noqa: F401
 from .ignored_processing import get_not_ignore_issue  # noqa: F401
 from .parser import (  # noqa: F401
     enumerate_file,
     get_ignore_files,
     get_ignore_lines,
+    get_ignored_blocks,
     parse_file,
 )
 from .visitor import PrintNodeVisitor, contains_print  # noqa: F401