+++ tests/testing_files/ignored_block/no_re_enable/ppl002.py
@@ -1,0 +1,11 @@
+# Standard imports
+from pprint import pprint
+
+
+def toto():
+    return 1 + 2
+
+
+# <py-printlinter disable PPL002>
+pprint(toto())
+pprint(toto() + toto())