+++ tests/testing_files/ignored_block/re_enable/ignore_nothing.py
@@ -1,0 +1,5 @@
+def toto():
+    return 1 + 2
+
+
+print(toto())