+++ tests/testing_files/ignored_block/re_enable/ppl001.py
@@ -1,0 +1,8 @@
+def toto():
+    return 1 + 2
+
+
+# <py-printlinter disable PPL001>
+print(toto())
+print(toto() + toto())
+# <py-printlinter enable PPL001>