+++ tests/testing_files/ignored_block/re_enable/ppl004.py
@@ -1,0 +1,12 @@
+# Standard imports
+from sys import stderr
+
+
+def toto():
+    return 1 + 2
+
+
+# <py-printlinter disable PPL004>
+stderr.write(toto())
+stderr.write(toto() + toto())
+# <py-printlinter enable PPL004>