+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Now we can ignore a block of code and don't re enable the linter.
+- Add a way to ignore a block of code with: `<py-printlinter disable PPLXXX>` to disable
+  linter and: `<py-printlinter enable PPLXXX>` to re enable it. We can also disable
+  `ALL` rules.
 - Add a configuration option to enable/disable colorized output: `color`.
 [Issue](https://github.com/lilisse/printlinter/issues/6).
 - Add a way to ignore next line with: `<py-printlinter disable-next PPLXXX>`.